Generate vanilla and profile .o files separately
Rather than modifying objects in-place with frank, the vanilla
and profile objects are generated separately to make them
easier to inspect.

Now that the profile and vanilla objects are created independently,
they can be compiled in parallel, reducing build time on many-core
systems.

This also obviates the need for a separate dummy epilogue dir and manual
list of epilogue-patched objects - we now apply frank indiscriminately
to all melee and sysdolphin sources.