Port `obj_files.mk` to `configure.py`
Note that this does not add splits or symbols, and so
does basically nothing to the `ninja` build process yet.

There are several orphaned data files which should not
exist, such as `baselib_shared_data_001`, but these are
products of erroneous splits and will be easy to iron out
later.