Fix include paths in `sysdolphin` (#1120)
Note that there is a bug in MWCC when using `-cwd source`. It will use the location of the current C file, not the header, therefore headers must be more specific in their paths.

For example, in `gobj.c`, `#include "gobj.h"` is fine because it resolves to `gobj.h` in the directory of `gobj.c`. But if you put `#include "jobj.h"` in `gobj.h`, and then `fighter.c` tries to include `gobj.h`, it will try to resolve `melee/ft/jobj.h` instead of the correct path.