Clobber dtk hash
Now we broke the caching. I usually just do `9r0` in vim to write a
bunch of zeros in the hash. You can also change it to `hash = ""`.
Anyway the hash no longer exists in the cache so Nix is gonna try to
fetch it.

The fetch is going to fail to match the garbage hash, but nix will
tell you what it got so you can copy/paste the correct value
into the hash field.