Add partial match for HSD_GObj_80391304
Almost matched, but there is a regswap in the final for loop.
This function handles initialization of the global GObj double-pointer
lists, and will not match if it is in the same file as those data
structures, so it is moved to the separate file "gobjinit.c". This
allows gobj.c to be linked in, as all the other functions were already
matching.