Port over THP from Pikmin 2
Other games with THP decomped seem to use later version which use a
global THPFileInfo pointer instead of passing it as an argument to each
private function. Making this change is enough to match or near-match
several functions, and many functions are close enough to confidently
name them in symbols.txt now.