Include main header first (#1786)
The "main" include (e.g. for the file fighter.c, `#include "fighter.h"`
should be the first include. This is because each .c file has its own
corresponding .h file, so including the main header first guarantees
that the .h file can be included "standalone" without requiring any
other transitive includes.

Clang-format gives the main include priority 0 by default, and doesn't
seem to allow this to be changed. So we should bump up the priority of
all the other include groups, so that 0 is the first group. The current
include order is maintained, except that the main header is now first.